Fix sidebar addCard action
