+++ app/containers/CardList.js
@@ -8,15 +8,6 @@ const { changeActiveBoard } = FILTERS;
 const { addCard } = CARD;
 const { toggleBoardsDialog, createBoard } = BOARDS;
 
-const NEW_CARD_TEMPLATE = {
-  title: 'New card',
-  text: `
-  ## Shortcuts
-  - ESC to discard changes
-  - Ctrl+Enter to save changes
-`,
-  editingText: '## Double click to edit card',
-};
 function mapDispatchToProps(dispatch) {
   return {
     toggleBoardsDialog: () => {
@@ -29,17 +20,7 @@ function mapDispatchToProps(dispatch) {
       dispatch(createBoard(board));
     },
     addCard: (activeBoard) => {
-      dispatch(
-        addCard({
-          title: NEW_CARD_TEMPLATE.title,
-          text: NEW_CARD_TEMPLATE.text,
-          board: activeBoard,
-          color: getRandomColor(),
-          isEditing: true,
-          editingTitle: NEW_CARD_TEMPLATE.title,
-          editingText: NEW_CARD_TEMPLATE.editingText,
-        }),
-      );
+      dispatch(addCard(activeBoard));
     },
   };
 }
